# segundo-parcial-inf133-api/app/static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Task Management API",
      "description": "API for managing tasks in a team",
      "version": "1.0.0"
    },
    "host": "localhost:5000",
    "basePath": "/api",
    "schemes": ["http"],
    "paths": {
      "/register": {
        "post": {
          "summary": "Register a new user",
          "consumes": ["application/json"],
          "parameters": [
            {
              "in": "body",
              "name": "user",
              "description": "User to register",
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "User created successfully"
            },
            "400": {
              "description": "User already exists"
            }
          }
        }
      },
      "/login": {
        "post": {
          "summary": "Login user",
          "consumes": ["application/json"],
          "parameters": [
            {
              "in": "body",
              "name": "user",
              "description": "User to login",
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "User logged in successfully"
            },
            "401": {
              "description": "Invalid credentials"
            }
          }
        }
      },
...
    },
    "definitions": {
      "User": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password"
          },
          "role": {
            "type": "string",
            "enum": ["admin", "member"],
            "default": "member"
          }
        }
      },
...
    }
  }